Распределяет утилитарные функции по файлам в папке src/utils
