+++ src/const.js
@@ -1,1 +1,1 @@
-export const BEFOREEND = `beforeend`;
+
