+++ src/utils/common.js
@@ -1,0 +1,15 @@
+export const getRandomInteger = (min = 0, max = 1) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const getRandomValueFromArray = (values) => values[getRandomInteger(0, values.length - 1)];
+
+export const getRandomDouble = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+
+  return (lower + Math.random() * (upper - lower + 1)).toFixed(1);
+};
