+++ src/utils/film.js
@@ -1,0 +1,15 @@
+export const getRandomDate = () => {
+  const start = new Date(1940, 10, 30);
+  const end = new Date();
+  const randomDate = new Date(start.getTime() + Math.random() * (end.getTime() - start.getTime()));
+
+  return randomDate.toLocaleString(`en-GB`, {day: `numeric`, month: `long`, year: `numeric`});
+};
+
+export const getHumanizeViewFromDuration = ({hours, minutes}) => {
+  let humanizeTimeString = ``;
+  humanizeTimeString += hours ? `${hours}h ` : ``;
+  humanizeTimeString += minutes ? `${minutes}m` : ``;
+
+  return humanizeTimeString;
+};
